backend/user-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
pydantic[email]==2.5.2
python-multipart==0.0.6
dnspython==2.4.2